CategoryService Updated

+++ src/E-Commerce.Application/Services/Baskets/IBasketService.cs
@@ -4,8 +4,8 @@ namespace E_Commerce.Application.Services.Baskets
 {
     public interface IBasketService
     {
-        public ValueTask<Basket> GetBasketAsync(int userTelegramId);
-        public ValueTask<Basket> UpdateBasketProductsAsync(Product product,int count, int userTelegramId);
+        public ValueTask<Basket> GetBasketAsync(long userTelegramId);
+        public ValueTask<Basket> UpdateBasketProductsAsync(Product product,int count, long userTelegramId);
         public ValueTask<Basket> UpdateBasketStatusAsync(int userTelegramId);
     }
 }
